apps/reco/price_model/requirements.txt
# Price Classification Model Requirements
lightgbm>=4.0
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
psycopg2-binary>=2.9
python-dotenv>=1.0
openpyxl>=3.1
xgboost>=2.0.0
